Update phantomcss.js
Just started using PhantomCSS, works great. Compliments!

A suggestion is to remove the sentences containing the word 'fail' in it. We log the test results to a logging server, with an e-mail message beeing sent in case the string "fail" is part of the output.

To make this work I had to change some lines.